requirements.txt
google-cloud-aiplatform>=1.38
python-dotenv>=1.0.1
pip>=24.2